+++ kelpie-f446/Core/UserUtils/UserTypes.h
@@ -22,8 +22,8 @@ typedef uint16_t pwm_t;
 
 typedef struct PiCommsMessage_t
 {
-	uint32_t messageId;
-	uint32_t dataLen;
+	uint16_t messageId;
+	uint16_t dataLen;
 	void *data;
 }PiCommsMessage_t;
